秦阳坤 实现订单项的增删改查，添加客户，服务员，地址等的模糊查询，修改客户的删除错误，添加类别，产品，订单等的删除异常判断功能

+++ src/main/java/com/briup/apps/ej/service/impl/ICategoryServiceImpl.java
@@ -10,17 +10,16 @@ import javax.annotation.Resource;
 import java.util.List;
 
 @Service
-public class ICategoryServiceImpl implements ICategoryService{
+public class ICategoryServiceImpl implements ICategoryService {
 
     @Resource
     private CategoryMapper categoryMapper;
 
     @Override
     public int saveOrUpdate(Category category) {
-        if(category.getId()==null){
-            return  categoryMapper.insert(category);
-        }
-        else {
+        if (category.getId() == null) {
+            return categoryMapper.insert(category);
+        } else {
             return categoryMapper.updateByPrimaryKey(category);
         }
     }
@@ -32,7 +31,13 @@ public class ICategoryServiceImpl implements ICategoryService{
     }
 
     @Override
-    public int deleteByPrimaryKey(Long id) {
-        return categoryMapper.deleteByPrimaryKey(id);
+    public int deleteByPrimaryKey(Long id) throws Exception {
+        Category category = categoryMapper.selectByPrimaryKey(id);
+
+        if (category == null) {
+            throw new Exception("要删除的用户不存在");
+        } else {
+            return categoryMapper.deleteByPrimaryKey(id);
+        }
     }
 }
